Make SYNC IMAGES(*) work by handling size of -1 in library.
libgfortran/ChangeLog:

	* caf_shared/sync.c (sync_table): Change size argument and index
	to int.
	* caf_shared/sync.h (sync_table): Adjust prototype.
	* caf_shared/wrapper.c (cas_sync_images): Add s argument to int,
	adjust call to sync_table.

gcc/testsuite/ChangeLog:

	* gfortran.dg/caf-shared/sync_images_1.f90: New test.
	* gfortran.dg/caf-shared/sync_images_2.f90: New test.

+++ libgfortran/caf_shared/sync.h
@@ -51,7 +51,7 @@ internal_proto (sync_iface_init);
 void sync_all (sync_iface *);
 internal_proto (sync_all);
 
-void sync_table (sync_iface *, int *, size_t);
+void sync_table (sync_iface *, int *, int);
 internal_proto (sync_table);
 
 #endif
